Feat: Finishing pipeline built adding generate-pipeline command

+++ ny_rides/contracts/yellow_taxi_contract.py
@@ -9,7 +9,7 @@ class YellowTaxiContract:
 
     COLUMN_TYPE_RULES = {
         "VendorID": "integer",
-        "passenger_count": "integer",
+        "passenger_count": "numeric",
         "total_amount": "numeric",
         "tpep_pickup_datetime": "datetime",
         "tpep_dropoff_datetime": "datetime",
